Refactor common methods to abstract LinuxCommand
With the addition of a toggleable arg for groups, there'll be quite a
bit of duplication between the LinuxUser and LinuxGroup classes, so this
takes the more generic methods into a new abstract class they extend.

This also has the benefit of greatly simplifying the constructors. Yay!
