+++ app/System/Users/LinuxUser.php
@@ -2,13 +2,10 @@
 
 namespace Servidor\System\Users;
 
-class LinuxUser
-{
-    /**
-     * @var array
-     */
-    private $args = [];
+use Servidor\System\LinuxCommand;
 
+class LinuxUser extends LinuxCommand
+{
     /**
      * @var string
      */
@@ -23,11 +20,6 @@ class LinuxUser
      * @var int
      */
     public $uid;
-
-    /**
-     * @var array
-     */
-    protected $original = [];
 
     /**
      * @var array
@@ -48,27 +40,13 @@ class LinuxUser
     {
         $this->name = $user['name'] ?? '';
 
-        foreach (['dir', 'gid', 'uid', 'shell'] as $key) {
-            if (isset($user[$key])) {
-                $this->$key = $user[$key];
-            }
-        }
+        $this->initArgs(['dir', 'gid', 'uid', 'shell'], $user);
 
         if ($loadGroups) {
             $this->loadGroups();
         }
 
-        $this->original = $this->toArray();
-    }
-
-    public function getArgs(): array
-    {
-        return $this->args;
-    }
-
-    public function getOriginal(string $key)
-    {
-        return $this->original[$key] ?? null;
+        $this->setOriginal();
     }
 
     public function setName(string $name): self
@@ -164,28 +142,6 @@ class LinuxUser
         return $this->toggleArg($enabled, '-U', '-N');
     }
 
-    public function toggleArg(bool $cond, string $on, string $off = ''): self
-    {
-        $keyOn = array_search($on, $this->args);
-        $keyOff = array_search($off, $this->args);
-
-        if (is_int($keyOn)) {
-            unset($this->args[$keyOn]);
-        }
-
-        if ('' != $keyOff && is_int($keyOff)) {
-            unset($this->args[$keyOff]);
-        }
-
-        $arg = $cond ? $on : $off;
-
-        if ('' != $arg) {
-            $this->args[] = $arg;
-        }
-
-        return $this;
-    }
-
     private function loadGroups(): void
     {
         $this->groups = [];
@@ -207,11 +163,6 @@ class LinuxUser
     public function isDirty(): bool
     {
         return count($this->args) > 0;
-    }
-
-    public function toArgs(): string
-    {
-        return implode(' ', $this->args);
     }
 
     public function toArray(): array
